backend/requirements.txt
Django==5.2.11
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
Pillow
reportlab
python-dateutil